+++ app/src/main/java/com/vandyke/sia/data/local/models/renter/Dir.kt
@@ -6,13 +6,12 @@ package com.vandyke.sia.data.local.models.renter
 
 import android.arch.persistence.room.Entity
 import android.arch.persistence.room.PrimaryKey
-import java.math.BigDecimal
 
 @Entity(tableName = "dirs")
 class Dir(
         @PrimaryKey
         override val path: String,
-        override val size: BigDecimal
+        override val size: Long
 ) : Node() {
     override val parent: String?
         get() = this.path.dirPathParent()
